chatbot/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
httpx>=0.24.0
aiofiles>=23.0.0
python-multipart>=0.0.5
jinja2>=3.0.0
